% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{classify_polarity}
\alias{classify_polarity}
\title{classify_polarity}
\usage{
classify_polarity(textColumns, algorithm = "bayes", pstrong = 0.5,
  pweak = 1, prior = 1, verbose = FALSE, ...)
}
\arguments{
\item{textColumns}{A \code{data.frame} of text documents listed one per row.}

\item{algorithm}{A \code{string} indicating whether to use the naive \code{bayes} algorithm or a simple \code{voter} algorithm.}

\item{pstrong}{A \code{numeric} specifying the probability that a strongly subjective term appears in the given text.}

\item{pweak}{A \code{numeric} specifying the probability that a weakly subjective term appears in the given text.}

\item{prior}{A \code{numeric} specifying the prior probability to use for the naive Bayes classifier.}

\item{verbose}{A \code{logical} specifying whether to print detailed output regarding the classification process.}

\item{\ldots}{Additional parameters to be passed into the \code{\link{create_matrix}} function.}
}
\description{
Classifies the polarity (e.g. positive or negative) of a set of texts using a naive Bayes classifier trained on Janyce Wiebe's \code{\link{lexicon}} lexicon. The words have been manually translated into Spanish and checked to make sure the polarity is right.
}
\examples{
# LOAD LIBRARY
library(sentimiento)
# DEFINE DOCUMENTS
documents <- c("I am very happy, excited, and optimistic.",
              "I am very scared, annoyed, and irritated.",
              "Iraq's political crisis entered its second week one step closer to the potential
				dissolution of the government, with a call for elections by a vital coalition partner
				and a suicide attack that extended the spate of violence that has followed the withdrawal
				of U.S. troops.",
              "With nightfall approaching, Los Angeles authorities are urging residents to keep their
				outdoor lights on as police and fire officials try to catch the person or people responsible
				for nearly 40 arson fires in the last three days.")

# CLASSIFY POLARITY
classify_polarity(documents,algorithm="bayes",verbose=TRUE)
}
\author{
Javier Sajuria \email{j.sajuria@ucl.ac.uk}
}
\keyword{sentiment}

